+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - discogs page: Added release hints on the Discogs edit page
 - popup: Added load hints link to the release on History page
 - popup: Added clear search button to the releases and history table
+- popup: Don't display the prepared discogs release csv data in the popup
 - core: Some code refactoring. Improved module loading
 
 ### Fixed